moved flatten in its own file
