+++ src/main/kotlin/io/konad/Result.kt
@@ -1,10 +1,7 @@
 package io.konad
 
 import io.konad.Maybe.Companion.maybe
-import io.konad.Maybe.Companion.nullable
-import io.konad.applicative.builders.ap
-import io.konad.applicative.builders.flatten
-import io.konad.applicative.builders.map
+import io.konad.applicative.flatten
 import io.konad.exceptions.ResultException
 import io.konad.hkt.ApplicativeFunctorKind
 import io.konad.hkt.FunctorKind
